refactor: replace endpoints to separate folder
