+++ src/endpoints/album/album.module.ts
@@ -3,8 +3,8 @@ import { AlbumService } from './album.service';
 import { AlbumController } from './album.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Album } from './models/album.entity';
-import { Track } from 'src/track/models/track.entity';
-import { FavoritesModule } from 'src/favorites/favorites.module';
+import { Track } from '../track/models/track.entity';
+import { FavoritesModule } from '../favorites/favorites.module';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Album, Track]), FavoritesModule],
