+++ src/endpoints/album/dto/update-album.dto.ts
