+++ src/endpoints/album/models/album.entity.ts
