+++ src/endpoints/artist/artist.controller.ts
