+++ src/endpoints/artist/dto/create-artist.dto.ts
