+++ src/endpoints/artist/dto/update-artist.dto.ts
