+++ src/endpoints/artist/models/artist.entity.ts
