+++ src/endpoints/favorites/favorites.controller.ts
