+++ src/endpoints/favorites/favorites.module.ts
@@ -3,9 +3,9 @@ import { FavoritesService } from './favorites.service';
 import { FavoritesController } from './favorites.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Favorites } from './models/favorites.entity';
-import { Artist } from 'src/artist/models/artist.entity';
-import { Album } from 'src/album/models/album.entity';
-import { Track } from 'src/track/models/track.entity';
+import { Artist } from '../artist/models/artist.entity';
+import { Album } from '../album/models/album.entity';
+import { Track } from '../track/models/track.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Favorites, Artist, Album, Track])],
