+++ src/endpoints/track/track.module.ts
@@ -3,7 +3,7 @@ import { TrackController } from './track.controller';
 import { TrackService } from './track.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Track } from './models/track.entity';
-import { FavoritesModule } from 'src/favorites/favorites.module';
+import { FavoritesModule } from '../favorites/favorites.module';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Track]), FavoritesModule],
