+++ src/endpoints/user/models/user.entity.ts
